both ex1 ex2

+++ module01/ex1/ft_garden_data.py
@@ -2,7 +2,7 @@
 
 
 class Plant:
-    def __init__(self, name: str, height: float, age: int) -> None:
+    def __init__(self, name: str, height: int, age: int) -> None:
         self.name = name
         self.height = height
         self.age = age
@@ -14,9 +14,9 @@ class Plant:
 if __name__ == "__main__":
     print("=== Garden Plant Registry ===")
 
-    rose = Plant("Rose", 25.0, 30)
-    sunflower = Plant("Sunflower", 80.0, 45)
-    cactus = Plant("Cactus", 15.0, 120)
+    rose = Plant("Rose", 25, 30)
+    sunflower = Plant("Sunflower", 80, 45)
+    cactus = Plant("Cactus", 15, 120)
 
     rose.show()
     sunflower.show()
